+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## [0.0.38]
+ - make publish less restrictive (allow changing language, author and any openiap settings)
+ - load repos at start, and update default repo list
 ## [0.0.37]
  - Cleanup commands
  - Add dynamic template selector for initializing projects